+++ CHANGELOG.md
@@ -10,11 +10,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add ECS mappings for more audit anomaly events. [#70](https://github.com/elastic/go-libaudit/pull/70)
 - Add `BacklogWaitTimeActual` status field, which is available since Linux 5.9 [#93](https://github.com/elastic/go-libaudit/pull/93/files)
 - Add ECS normalizations for `TIME_ADJNTPVAL` and `TIME_INJOFFSET`. [#98](https://github.com/elastic/go-libaudit/pull/98)
+- Add support for exe filters in exclude rules (e.g. `-a exclude,always -F exe=/bin/ls`). [#97](https://github.com/elastic/go-libaudit/pull/97)
 
 ### Changed
 
 - Update syscall, arches, and audit msg type tables for Linux 5.16. [#96](https://github.com/elastic/go-libaudit/pull/96)
 - Go 1.16 or newer is required because the project uses the embed package. [#104](https://github.com/elastic/go-libaudit/pull/104)
+- Fixed error messages from `AddRule()` in the audit client. [#103](https://github.com/elastic/go-libaudit/pull/103)
 
 ### Removed